+++ gps/src/main/java/com/google/android/things/contrib/driver/gps/NmeaParser.java
@@ -16,8 +16,14 @@
 
 package com.google.android.things.contrib.driver.gps;
 
+import android.location.GnssStatus;
+import android.location.Location;
+import android.location.LocationManager;
 import android.text.format.DateFormat;
 import android.util.Log;
+import android.util.SparseArray;
+
+import com.google.android.things.userdriver.location.GnssStatusBuilder;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -41,16 +47,19 @@ import java.util.TimeZone;
 
     // NMEA Message Types
     private static final String GGA = "GPGGA";
+    private static final String GSV = "GPGSV";
     private static final String GLL = "GPGLL";
     private static final String RMC = "GPRMC";
-    private static final String VTG = "GPVTG";
 
     private Calendar mTimestampCalendar;
     private GpsModuleCallback mGpsModuleCallback;
 
-    /*package*/ NmeaParser() {
+    private float mGpsAccuracy;
+
+    /*package*/ NmeaParser(float gpsAccuracy) {
         // Initialize timestamp calendar to current time
         mTimestampCalendar = Calendar.getInstance(TimeZone.getTimeZone("UTC"));
+        mGpsAccuracy = gpsAccuracy;
     }
 
     /*package*/ void setGpsModuleCallback(GpsModuleCallback callback) {
@@ -78,21 +87,24 @@ import java.util.TimeZone;
             return;
         }
 
-        // Parse the message based on type
+        // Report the raw, validated message
         String nmea = new String(message, 0, index);
+        postRaw(nmea);
+
+        // Parse the message based on type
         String[] tokens = nmea.split(DELIMITER);
         switch (tokens[0]) {
             case GGA:
                 handleFixInformation(tokens);
+                break;
+            case GSV:
+                handleSatelliteData(tokens);
                 break;
             case GLL:
                 handleLatLngData(tokens);
                 break;
             case RMC:
                 handleRecommendedMinimum(tokens);
-                break;
-            case VTG:
-                handleVelocityTrackData(tokens);
                 break;
             default:
                 // Ignore the message
@@ -121,7 +133,7 @@ import java.util.TimeZone;
 
         int checkSum = convertAsciiByte(message[index+1], message[index+2]);
         if (messageSum != checkSum) {
-            Log.w(TAG, "Invalid checksum (" + messageSum + "), expected " + checkSum);;
+            Log.w(TAG, "Invalid checksum (" + messageSum + "), expected " + checkSum);
             return -1;
         }
 
@@ -139,9 +151,6 @@ import java.util.TimeZone;
         }
 
         int quality = Integer.parseInt(nmea[6]);
-        int satelliteCount = Integer.parseInt(nmea[7]);
-        postSatelliteStatus(quality > 0, satelliteCount);
-
         if (quality < 1) {
             // No valid fix
             return;
@@ -152,7 +161,53 @@ import java.util.TimeZone;
         double longitude = parseCoordinate(nmea[4], nmea[5]);
         double altitude = parseDistance(nmea[9], nmea[10]);
         double seaLevel = parseDistance(nmea[11], nmea[12]);
-        postPosition(timestamp, latitude, longitude, altitude - seaLevel);
+        postLocation(timestamp, latitude, longitude, altitude - seaLevel, -1, -1);
+    }
+
+    /** Intermediate representation of satellite data */
+    private SparseArray<Satellite> mSatellites = new SparseArray<>();
+    private static class Satellite {
+        int svid;
+        float elevation;
+        float azimuth;
+        float snr;
+    }
+    /**
+     * Parse the contents of a GPGSV sentence
+     * @param nmea Sentence tokens
+     */
+    private void handleSatelliteData(String[] nmea) {
+        if (nmea.length < 19) {
+            Log.w(TAG, "Invalid GSV Message");
+            return;
+        }
+
+        int satelliteCount = Integer.parseInt(nmea[3]);
+        if (satelliteCount < 1) {
+            // No valid fix
+            return;
+        }
+
+        // Parse the satellites in this message
+        for (int i = 4; i < 19; i += 4) {
+            if (!nmea[i].isEmpty()) {
+                Satellite sat = new Satellite();
+                sat.svid = Integer.parseInt(nmea[i]);
+                sat.elevation = parseTrackAngle(nmea[i+1]);
+                sat.azimuth = parseTrackAngle(nmea[i+2]);
+                sat.snr = parseSignal(nmea[i+3]);
+
+                mSatellites.put(sat.svid, sat);
+            }
+        }
+
+        if (mSatellites.size() < satelliteCount) {
+            // We haven't yet received all the satellite data
+            return;
+        }
+
+        postSatelliteStatus(mSatellites);
+        mSatellites.clear();
     }
 
     /**
@@ -174,7 +229,7 @@ import java.util.TimeZone;
         long timestamp = getUpdatedTimestamp(nmea[5], null);
         double latitude = parseCoordinate(nmea[1], nmea[2]);
         double longitude = parseCoordinate(nmea[3], nmea[4]);
-        postPosition(timestamp, latitude, longitude, -1);
+        postLocation(timestamp, latitude, longitude, -1, -1, -1);
     }
 
     /**
@@ -195,33 +250,17 @@ import java.util.TimeZone;
 
         long timestamp = getUpdatedTimestamp(nmea[1], nmea[9]);
         postTime(timestamp);
+
         double latitude = parseCoordinate(nmea[3], nmea[4]);
         double longitude = parseCoordinate(nmea[5], nmea[6]);
-        postPosition(timestamp, latitude, longitude, -1);
-
         float speed = parseSpeed(nmea[7], "N");
         float bearing = parseTrackAngle(nmea[8]);
-        postSpeed(speed, bearing);
+        postLocation(timestamp, latitude, longitude, -1, speed, bearing);
     }
 
-    /**
-     * Parse the contents of a GPVTG sentence
-     * @param nmea Sentence tokens
-     */
-    private void handleVelocityTrackData(String[] nmea) {
-        if (nmea.length < 9) {
-            Log.w(TAG, "Invalid VTG Message");
-            return;
-        }
-
-        float bearing = parseTrackAngle(nmea[1]);
-        float speed = parseSpeed(nmea[5], nmea[6]);
-        postSpeed(speed, bearing);
-    }
-
-    private void postSatelliteStatus(boolean active, int satellites) {
+    private void postRaw(String rawNmea) {
         if (mGpsModuleCallback != null) {
-            mGpsModuleCallback.onGpsSatelliteStatus(active, satellites);
+            mGpsModuleCallback.onNmeaMessage(rawNmea);
         }
     }
 
@@ -231,25 +270,56 @@ import java.util.TimeZone;
         }
     }
 
-    private void postPosition(long timestamp, double latitude, double longitude, double altitude) {
+    private void postSatelliteStatus(SparseArray<Satellite> satellites) {
         if (mGpsModuleCallback != null) {
-            mGpsModuleCallback.onGpsPositionUpdate(timestamp, latitude, longitude, altitude);
+            GnssStatusBuilder builder = new GnssStatusBuilder(satellites.size())
+                    .setFlags(0, GnssStatusBuilder.GNSS_SV_FLAGS_NONE)
+                    .setConstellation(0, GnssStatus.CONSTELLATION_GPS);
+
+            for (int i = 0; i < satellites.size(); i++) {
+                Satellite sat = satellites.valueAt(i);
+                builder.setSvid(i, sat.svid)
+                        .setElevation(i, sat.elevation)
+                        .setAzimuth(i, sat.azimuth)
+                        .setCn0DbHz(i, sat.snr);
+            }
+
+            mGpsModuleCallback.onGpsSatelliteStatus(builder.build());
         }
     }
 
-    private void postSpeed(float speed, float bearing) {
+    private void postLocation(long timestamp, double latitude, double longitude, double altitude, float speed, float bearing) {
         if (mGpsModuleCallback != null) {
-            mGpsModuleCallback.onGpsSpeedUpdate(speed, bearing);
+            Location location = new Location(LocationManager.GPS_PROVIDER);
+            // We cannot compute accuracy from NMEA data alone.
+            // Assume that a valid fix has the quoted accuracy of the module.
+            // Framework requires accuracy in DRMS.
+            location.setAccuracy(mGpsAccuracy * 1.2f);
+            location.setTime(timestamp);
+
+            location.setLatitude(latitude);
+            location.setLongitude(longitude);
+            if (altitude != -1) {
+                location.setAltitude(altitude);
+            }
+            if (speed != -1) {
+                location.setSpeed(speed);
+            }
+            if (bearing != -1) {
+                location.setBearing(bearing);
+            }
+
+            mGpsModuleCallback.onGpsLocationUpdate(location);
         }
     }
 
+    private static final SimpleDateFormat FORMAT = new SimpleDateFormat("ddMMyyHHmmss", Locale.US);
     /**
      * Apply the NMEA time parameters and get an updated timestamp
      * @param timeString NMEA fix time string
      * @param dateString NMEA date string, optional
      * @return Epoch timestamp.
      */
-    private static final SimpleDateFormat FORMAT = new SimpleDateFormat("ddMMyyHHmmss", Locale.US);
     private long getUpdatedTimestamp(String timeString, String dateString) {
         if (timeString.length() < 6) {
             // Invalid time
@@ -319,7 +389,25 @@ import java.util.TimeZone;
     }
 
     /**
-     * Combine an NMEA angle string into a decimal value in true degrees.
+     * Convert a Signal-Noise Ratio string into a decimal value in dB.
+     * @param signalString Integer string of the SNR.
+     * @return Decimal value, or -1 if null
+     */
+    private float parseSignal(String signalString) {
+        if (signalString.isEmpty()) {
+            return -1;
+        }
+
+        try {
+            return Float.parseFloat(signalString);
+        } catch (NumberFormatException e) {
+            // Invalid value
+            return -1;
+        }
+    }
+
+    /**
+     * Convert an NMEA angle string into a decimal value in true degrees.
      * @param angleString Decimal string of the angle.
      * @return Decimal value converted to true degrees.
      */
@@ -366,14 +454,14 @@ import java.util.TimeZone;
         }
     }
 
+    private static final float KNOTS = 0.514444f;
+    private static final float KPH = 0.277778f;
     /**
      * Combine an NMEA speed tuple into a decimal value in m/s.
      * @param speedString Decimal string of the speed.
      * @param unit Unit of measure constant.
      * @return Decimal value converted to m/s.
      */
-    private static final float KNOTS = 0.514444f;
-    private static final float KPH = 0.277778f;
     private float parseSpeed(String speedString, String unit) {
         if (speedString.isEmpty() || unit.isEmpty()) {
             // No data